pyproject.toml
[project]
name = "hands-onllmops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.24",
    "python-dotenv>=1.1.0",
    "uqlm>=0.1.6",
]
